[Android] The mapping of permissions to AndroidManifest.xml.
  This feature provides the support of mapping the manifest permissisons
to AndroidManifest.xml for the web app user in packaging tool. It provides
the function interface of mapping the permissions from manifest.json to
AndroidManifest.xml based on the field parser and mapping table. It suppports
the mapping of permission fields both defined in Crosswalk and supported by
Android Manifest specification. The test cases are added accordingly.

BUG=https://crosswalk-project.org/jira/browse/XWALK-651